// Project_Apex/frontend/src/api/models/NotificationsPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { NotificationPublic } from "./NotificationPublic";

/**
 * Public schema for list of notifications
 */
export type NotificationsPublic = {
    data: Array<NotificationPublic>;
    count: number;
};